Change from Integer to Double in Current model

+++ app/src/main/java/pl/piotrb/weatherapp/model/oncecallapi/Current.java
@@ -26,7 +26,7 @@ public class Current {
     private Integer clouds;
     private Integer visibility;
     @SerializedName("wind_speed")
-    private Integer windSpeed;
+    private Double windSpeed;
     @SerializedName("wind_deg")
     private Integer windDeg;
     private List<Weather> weather = null;
